refactor: prepare moving ui package components to www app.
