+++ apps/www/app/components/index.ts
@@ -1,0 +1,1 @@
+export * from './theme-menu/theme-menu';
